requirements.txt
pandas
scholarly
requests
httpx<=0.27.0
python-dotenv
sentence-transformers
beautifulsoup4
PyGithub
